Break client up into logical blocks
